Minor corrections in RSA order functionality for PDPP
It was made some corrections in the functions which enables the RSA order functionalities and ProtectionDisjointYEN function. The PDPP-RSA/SAR is running properly and its PB values is close.  The PDPP-SAR simulation time is higher than RSA. It is necessary optimize the code to try reducing the PDPP-SAR simulation time.